+++ composeApp/src/commonMain/kotlin/domain/Orders/OrderRepository.kt
@@ -4,5 +4,5 @@ import com.badoo.reaktive.single.Single
 import domain.Models.OrderListDataResponse
 
 interface OrderRepository {
-    suspend fun getUserOrders(userId: Int, nextPage: Int = 1): Single<OrderListDataResponse>
+    suspend fun getUserOrders(userId: Long, nextPage: Int = 1): Single<OrderListDataResponse>
 }
