+++ composeApp/src/commonMain/kotlin/domain/Orders/OrderRequestContract.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetOrderRequest(@SerialName("user_id") val userId: Int)
+data class GetOrderRequest(@SerialName("user_id") val userId: Long)
