+++ composeApp/src/commonMain/kotlin/domain/Products/ProductRepository.kt
@@ -5,7 +5,7 @@ import domain.Models.ProductListDataResponse
 import domain.Models.ServerResponse
 
 interface ProductRepository {
-    suspend fun getAllProducts(vendorId: Int, nextPage: Int = 1): Single<ProductListDataResponse>
-    suspend fun searchProducts(vendorId: Int, searchQuery: String, nextPage: Int = 1): Single<ProductListDataResponse>
-    suspend fun createOrder(vendorId: Int,userId: Int,orderReference: Int,deliveryMethod: String,paymentMethod: String,orderItems: ArrayList<OrderItemRequest>): Single<ServerResponse>
+    suspend fun getAllProducts(vendorId: Long, nextPage: Int = 1): Single<ProductListDataResponse>
+    suspend fun searchProducts(vendorId: Long, searchQuery: String, nextPage: Int = 1): Single<ProductListDataResponse>
+    suspend fun createOrder(vendorId: Long,userId: Long,orderReference: Int,deliveryMethod: String,paymentMethod: String,orderItems: ArrayList<OrderItemRequest>): Single<ServerResponse>
 }
