+++ composeApp/src/commonMain/kotlin/domain/Products/ProductRequestContract.kt
@@ -7,19 +7,19 @@ import kotlinx.serialization.Serializable
 
 
 @Serializable
-data class GetAllProductsRequest(@SerialName("vendorId") val vendorId: Int)
+data class GetAllProductsRequest(@SerialName("vendorId") val vendorId: Long)
 
 @Serializable
-data class SearchProductRequest(@SerialName("vendorId") val vendorId: Int,
+data class SearchProductRequest(@SerialName("vendorId") val vendorId: Long,
                                         @SerialName("searchQuery") val searchQuery: String)
 
 @Serializable
-data class CreateOrderRequest(@SerialName("vendor_id") val vendorId: Int, @SerialName("user_id") val userId: Int,
+data class CreateOrderRequest(@SerialName("vendor_id") val vendorId: Long, @SerialName("user_id") val userId: Long,
                               @SerialName("orderReference") val orderReference: Int, @SerialName("deliveryMethod") val deliveryMethod: String,
                               @SerialName("paymentMethod") val paymentMethod: String, @SerialName("orderItem_Array") val orderItems: ArrayList<OrderItemRequest>)
 
 @Serializable
 data class OrderItemRequest(@SerialName("orderReference") val orderReference: Int,
-                            @SerialName("user_id") val userId: Int,
+                            @SerialName("user_id") val userId: Long,
                             @SerialName("product_id") val productId: Int,
                             @SerialName("itemCount") val itemCount: Int)
