+++ composeApp/src/commonMain/kotlin/presentation/appointments/AppointmentContract.kt
@@ -21,8 +21,8 @@ interface AppointmentContract {
 
     abstract class Presenter {
         abstract fun registerUIContract(view: View?)
-        abstract fun getUserAppointments(userId: Int)
-        abstract fun getMoreAppointments(userId: Int, nextPage: Int = 1)
+        abstract fun getUserAppointments(userId: Long)
+        abstract fun getMoreAppointments(userId: Long, nextPage: Int = 1)
         abstract fun postponeAppointment(appointment: Appointment, newAppointmentTime: Int,  day: Int, month: Int, year: Int)
         abstract fun deleteAppointment(appointmentId: Int)
         abstract fun joinMeeting(customParticipantId: String, presetName: String, meetingId: String)
